# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Baby Health Application API
  version: 1.0.0
  description: Multi-baby health record API (India). Reference values are returned only with provenance and only when their release gate is CLEARED.
    The API never returns a computed milk target for healthy infants.
  contact:
    name: Product owner
servers:
- url: https://api.example.in
  description: Production (India region) - placeholder
security:
- bearerAuth: []
...
paths:
...
  /v1/babies/{babyId}/measurements:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    get:
      summary: List weights
      tags:
      - Weights
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Measurement'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    post:
      summary: Create weights record
      tags:
      - Weights
      x-authorization: OWNER, GUARDIAN, CAREGIVER
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Measurement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '409':
          $ref: '#/components/responses/DuplicateSuspected'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementCreate'
  /v1/babies/{babyId}/measurements/{id}:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    - $ref: '#/components/parameters/Id'
    get:
      summary: Get weights record
      tags:
      - Weights
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Measurement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    patch:
      summary: Update weights record
      tags:
      - Weights
      x-authorization: OWNER, GUARDIAN, CAREGIVER
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Measurement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '412':
          $ref: '#/components/responses/VersionMismatch'
        '422':
          $ref: '#/components/responses/RuleViolation'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementPatch'
    delete:
      summary: Soft-delete weights record
      tags:
      - Weights
      x-authorization: OWNER, GUARDIAN, CAREGIVER
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '412':
          $ref: '#/components/responses/VersionMismatch'
  /v1/babies/{babyId}/growth:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    - name: indicator
      in: query
      required: true
      schema:
        enum:
        - WEIGHT_FOR_AGE
        - LENGTH_HEIGHT_FOR_AGE
        - WEIGHT_FOR_LENGTH
        - WEIGHT_FOR_HEIGHT
        - BMI_FOR_AGE
        - HEAD_CIRCUMFERENCE_FOR_AGE
        - WEIGHT_VELOCITY
        - LENGTH_VELOCITY
        - HEAD_CIRCUMFERENCE_VELOCITY
    get:
      summary: Growth reference for an indicator
      tags:
      - Growth
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrowthResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '422':
          $ref: '#/components/responses/RuleViolation'
      description: 422 GROWTH_REFERENCE_UNAVAILABLE when sex NOT_STATED, age > 60 months, or dataset not imported.
  /v1/growth/datasets:
    get:
      summary: List imported WHO datasets
      tags:
      - Growth
      x-authorization: Any authenticated user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token <= 15 min; authorization resolved per baby via membership + database row-level security.
  parameters:
    BabyId:
      name: babyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resources of babies without membership return 404.
...
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    From:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      schema:
        type: string
        format: date-time
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum:
        - occurred_at
        - -occurred_at
        - created_at
        - -created_at
        default: -occurred_at
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
      description: Record version
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/400
            title: Validation failed
            status: 400
    Unauthenticated:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/401
            title: Authentication required
            status: 401
    Forbidden:
      description: Role does not permit this action
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/403
            title: Role does not permit this action
            status: 403
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/404
            title: Not found
            status: 404
    DuplicateSuspected:
      description: Possible duplicate; resend with force=true to save
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/409
            title: Possible duplicate; resend with force=true to save
            status: 409
...
    VersionMismatch:
      description: Record changed; refetch
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/412
            title: Record changed; refetch
            status: 412
...
    RuleViolation:
      description: Business rule violated (e.g. quantity sent for DIRECT_BREASTFEEDING)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/422
            title: Business rule violated (e.g. quantity sent for DIRECT_BREASTFEEDING)
            status: 422
    RateLimited:
      description: Too many requests (see Retry-After)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/429
            title: Too many requests (see Retry-After)
            status: 429
    Internal:
      description: Internal error (support code provided)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/500
            title: Internal error (support code provided)
            status: 500
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        support_code:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
    Audit:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        version:
          type: integer
    Provenance:
      type: object
      required:
      - source_id
      - organization
      - document_title
      - source_url
      - page_or_section
      - verification_status
      properties:
        source_id:
          type: string
        organization:
          type: string
        document_title:
          type: string
        document_type:
          type: string
        country:
          type: string
        population:
          type: string
        clinical_context:
          type: string
        source_url:
          type: string
        publication_date:
          type: string
        version:
          type: string
        page_or_section:
          type: string
        retrieved_at:
          type: string
        verified_at:
          type: string
        evidence_level:
          type: string
        verification_status:
          type: string
        notes:
          type: string
...
    MeasurementCreate:
      type: object
      required:
      - measured_at
      - measured_tz
      - measurement_source
      minProperties: 4
      description: At least one of weight_kg, length_cm, head_circumference_cm
      anyOf:
      - required:
        - weight_kg
      - required:
        - length_cm
      - required:
        - head_circumference_cm
      properties:
        measured_at:
          type: string
          format: date-time
        measured_tz:
          type: string
        weight_kg:
          type:
          - number
          - 'null'
          exclusiveMinimum: 0
        length_cm:
          type:
          - number
          - 'null'
        length_position:
          enum:
          - RECUMBENT
          - STANDING
          - null
        head_circumference_cm:
          type:
          - number
          - 'null'
        measurement_source:
          enum:
          - HOME_SCALE
          - CLINIC
          - HOSPITAL
          - ANGANWADI
          - OTHER
        notes:
          type:
          - string
          - 'null'
    MeasurementPatch:
      type: object
      properties:
        measured_at:
          type: string
          format: date-time
        measured_tz:
          type: string
        weight_kg:
          type:
          - number
          - 'null'
          exclusiveMinimum: 0
        length_cm:
          type:
          - number
          - 'null'
        length_position:
          enum:
          - RECUMBENT
          - STANDING
          - null
        head_circumference_cm:
          type:
          - number
          - 'null'
        measurement_source:
          enum:
          - HOME_SCALE
          - CLINIC
          - HOSPITAL
          - ANGANWADI
          - OTHER
        notes:
          type:
          - string
          - 'null'
    Measurement:
      allOf:
      - $ref: '#/components/schemas/MeasurementCreate'
      - $ref: '#/components/schemas/Audit'
      - type: object
        properties:
          id:
            type: string
            format: uuid
          change_since_previous_g:
            type:
            - number
            - 'null'
          days_since_previous:
            type:
            - integer
            - 'null'
          g_per_day:
            type:
            - number
            - 'null'
    GrowthPoint:
      type: object
      properties:
        measurement_id:
          type: string
          format: uuid
        x:
          type: number
        value:
          type: number
        z_score:
          type:
          - number
          - 'null'
        percentile:
          type:
          - number
          - 'null'
        label:
          const: Growth reference
    GrowthResult:
      type: object
      properties:
        indicator:
          enum:
          - WEIGHT_FOR_AGE
          - LENGTH_HEIGHT_FOR_AGE
          - WEIGHT_FOR_LENGTH
          - WEIGHT_FOR_HEIGHT
          - BMI_FOR_AGE
          - HEAD_CIRCUMFERENCE_FOR_AGE
          - WEIGHT_VELOCITY
          - LENGTH_VELOCITY
          - HEAD_CIRCUMFERENCE_VELOCITY
        dataset_id:
          type: string
        bands:
          type: array
          items:
            type: object
          description: Copied from imported WHO dataset
        points:
          type: array
          items:
            $ref: '#/components/schemas/GrowthPoint'
        caveats:
          type: array
          items:
            type: string
        provenance:
          $ref: '#/components/schemas/Provenance'
...